adapted to citygml4j 2.1

+++ src/org/citygml4j/ade_xjc/Util.java
@@ -11,29 +11,29 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
 public class Util {
-	
+
 	public static BigInteger dir2md5(File file, BigInteger md5) throws NoSuchAlgorithmException, FileNotFoundException {
 		md5 = string2md5(file.getName(), md5); 
-		
+
 		if (file.isDirectory()) {
 			File[] children = file.listFiles(new FileFilter() {				
 				public boolean accept(File pathname) {
 					if (pathname.isDirectory())
 						return true;
-					
+
 					if (pathname.getName().endsWith("xsd") || pathname.getName().endsWith("xjb"))
 						return true;
-					
+
 					return false;
 				}
 			});
-			
+
 			for (File next : children)
 				md5 = dir2md5(next, md5);
-			
+
 		} else
 			md5 = file2md5(file.getAbsolutePath(), md5);
-		
+
 		return md5;
 	}
 
@@ -48,7 +48,7 @@ public class Util {
 				digest.update(buffer, 0, read);
 
 			byte[] md5sum = digest.digest();
-			
+
 			return md5.xor(new BigInteger(1, md5sum));
 		} catch(IOException e) {
 			throw new RuntimeException("Unable to process file for MD5", e);
@@ -65,23 +65,24 @@ public class Util {
 	private static BigInteger string2md5(String input, BigInteger md5) throws NoSuchAlgorithmException {
 		MessageDigest digest = MessageDigest.getInstance("MD5");	
 		byte[] md5sum = digest.digest(input.getBytes());
-		
+
 		return md5.xor(new BigInteger(1, md5sum));
 	}
-	
-	public static void rmdir(File folder) {
+
+	public static void rmdir(File folder, boolean force) {
 		if (!folder.exists())
 			return;
-		
+
 		File[] children = folder.listFiles();
-		
+
 		for (File file : children) {
 			if (file.isDirectory())
-				rmdir(file);
-			
-			file.delete();
+				rmdir(file, force);
+
+			if (force)
+				file.delete();
 		}
-		
+
 		folder.delete();
 	}
 }
